+++ R/split.R
@@ -30,7 +30,10 @@ coloc.detail <- function(dataset1, dataset2, MAF=NULL,
     
   df1 <- as.data.table(process.dataset(d=dataset1, suffix="df1"))[!is.na(lABF.df1),]
   df2 <- as.data.table(process.dataset(d=dataset2, suffix="df2"))[!is.na(lABF.df2),]
-  df <- merge(df1,df2,by="snp")
+    if("position" %in% names(df1) && "position" %in% names(df2))
+        df <- merge(df1,df2,by=c("snp","position"))
+    else
+        df <- merge(df1,df2,by="snp")
 
    if(!nrow(df))
     stop("dataset1 and dataset2 should contain the same snps in the same order, or should contain snp names through which the common snps can be identified")
@@ -38,7 +41,7 @@ coloc.detail <- function(dataset1, dataset2, MAF=NULL,
   df$lABF.h4 <- with(df, lABF.df1 + lABF.df2)
   ## add SNP.PP.H4 - post prob that each SNP is THE causal variant for a shared signal
   my.denom.log.abf <- logsum(df$lABF.h4)
-  df$H4 <- exp(df$lABF.h4 - my.denom.log.abf)
+  df$SNP.PP.H4 <- exp(df$lABF.h4 - my.denom.log.abf)
     df3 <- as.data.table(expand.grid(snp1=df$snp,snp2=df$snp))
     df3 <- merge(df3,df1[,.(snp,lABF.df1)],by.x="snp1",by.y="snp")
     df3 <- merge(df3,df2[,.(snp,lABF.df2)],by.x="snp2",by.y="snp")
@@ -52,8 +55,8 @@ coloc.detail <- function(dataset1, dataset2, MAF=NULL,
   results <- c(nsnps=common.snps, pp.abf)
   
     list(summary=results,
-         df=df,
-         df3=df3,
+         results=df,
+         results.H3=df3,
          priors=c(p1=p1,p2=p2,p12=p12)
          )
 }
@@ -145,7 +148,8 @@ est_cond <- function(x,LD,N,YY,sigsnps,cc=FALSE) {
     vb2 <- c(Sc) * ( D[use] - diag(XX21 %*% solve(XX1) %*% t(XX21))) / (D[use]^2)
     vb2 <- c(vb2)
     ## na <- apply(abs(LD[sigsnps,use,drop=FALSE])==1,2,any)
-    data.table(snp=x$snp[use],beta=c(b2),varbeta=vb2)
+    rbind(data.table(snp=x$snp[use],beta=c(b2),varbeta=vb2),
+          data.table(snp=x$snp[!use],beta=rep(0,sum(!use)),varbeta=x$varbeta[!use]))
     ## if(any(na))
     ##     ret[na,c("beta","varbeta"):=list(NA,NA)]
 }
@@ -245,6 +249,7 @@ finemap.signals <- function(D,LD=D$LD,aligned=FALSE,
                                   sigsnps=NULL,
                                   pthr=1e-6,
                                   maxhits=3) {
+    check.dataset(D)
     zthr=qnorm(pthr/2,lower.tail = FALSE)
     ## make nicer data table
     ## x <- as.data.table(D[c("beta","varbeta","snp","MAF")])
@@ -260,6 +265,8 @@ finemap.signals <- function(D,LD=D$LD,aligned=FALSE,
        }
     ## check LD in matching order
     LD <- LD[D$snp,D$snp]
+    if(ncol(LD)<2)
+        stop("<2 SNPs in D found in LD matrix - do the SNP names match?")
     hits <- NULL
     while(length(hits)<maxhits) {
         newhit=if(method=="mask") {
@@ -269,7 +276,7 @@ finemap.signals <- function(D,LD=D$LD,aligned=FALSE,
                             YY, # sum(Y^2)
                             sigsnps=names(hits))
                }
-        if(is.null(newhit) || abs(newhit) < zthr )
+        if(is.null(newhit) || !length(newhit) || abs(newhit) < zthr )
             break
         hits <- c(hits,newhit)
     }
@@ -284,18 +291,20 @@ finemap.signals <- function(D,LD=D$LD,aligned=FALSE,
 ##' @param hits1 lead snps for trait 1. If length > 1, will use masking
 ##' @param hits2 lead snps for trait 2. If length > 1, will use masking
 ##' @param LD named LD matrix (for masking)
+##' @param LD1 named LD matrix (for masking) for trait 1 only
+##' @param LD2 named LD matrix (for masking) for trait 2 only
 ##' @param r2thr r2 threshold at which to mask
 ##' @inheritParams coloc.abf
 ##' @return data.table of coloc results
 ##' @author Chris Wallace
-coloc.process <- function(result,hits1=NULL,hits2=NULL,LD=NULL,r2thr=0.01,p1=1e-4,p2=1e-4,p12=1e-6,LD1=LD,LD2=LD) {
+coloc.process <- function(obj,hits1=NULL,hits2=NULL,LD=NULL,r2thr=0.01,p1=1e-4,p2=1e-4,p12=1e-6,LD1=LD,LD2=LD,detail=FALSE) {
     ## which format?
     ## if("results" %in% names(result)) { # for legacy objects
-    ##     result$df <- result$results
-    ##     setnames(result$df3,c("snpA","snpB"),c("snp1","snp2"))
+    ##     obj$results <- obj$results
+    ##     setnames(obj$results3,c("snpA","snpB"),c("snp1","snp2"))
     ## }
-    setnames(result$df,paste0("lABF.h",c(1,2,4)),paste0("lbf",c(1,2,4)),skip_absent=TRUE)
-    setnames(result$df3,paste0("lABF.h",3),paste0("lbf",3),skip_absent=TRUE)
+    setnames(obj$results,paste0("lABF.h",c(1,2,4)),paste0("lbf",c(1,2,4)),skip_absent=TRUE)
+    setnames(obj$results.H3,paste0("lABF.h",3),paste0("lbf",3),skip_absent=TRUE)
     ## overall coloc on a given set of snp results
     .f <- function(df,df3) {
         lH0.abf <- 0
@@ -321,10 +330,13 @@ coloc.process <- function(result,hits1=NULL,hits2=NULL,LD=NULL,r2thr=0.01,p1=1e-
     
     ## one signal per trait
     if(length(hits1)<=1 & length(hits2)<=1) {
-        tmp <- .f(result$df,result$df3)
+        tmp <- .f(obj$results,obj$results.H3)
         tmp$hit1=hits1
         tmp$hit2=hits2
-        return(tmp)
+        ## setnames(obj$results,"H4","SNP.PP.H4")
+        return(list(summary=tmp,
+                    results=obj$results,
+                    priors=c(p1,p2,p12)))
     }
 
     ## otherwise mask all but one signal from each trait with  >1 signal
@@ -332,7 +344,11 @@ coloc.process <- function(result,hits1=NULL,hits2=NULL,LD=NULL,r2thr=0.01,p1=1e-
     ldfriends2 <-  LD2[setdiff(unique(c(hits2)),""),,drop=FALSE]^2
     todo <- expand.grid(i=seq_along(hits1),j=seq_along(hits2))
     todo <- t(todo) #[i<=j,])
-    ret <- lapply(1:ncol(todo), function(r) {
+    newresult <- obj$results[,intersect(c("snp","position"),
+                                      colnames(obj$results)),
+                           drop=FALSE,with=FALSE]
+    ret <- vector("list",ncol(todo))
+    for(r in 1:ncol(todo)) {
         i <- todo[1,r]
         j <- todo[2,r]
         message(r)
@@ -352,13 +368,19 @@ coloc.process <- function(result,hits1=NULL,hits2=NULL,LD=NULL,r2thr=0.01,p1=1e-
         ## drop2 <- colnames(ldfriends)[ld2 > r2thr]
         dropsnps <- unique(c(drop1,drop2))
 
-        message("dropping ",length(dropsnps),"/",nrow(result$df)," : ",length(drop1)," + ",length(drop2))
-        if(length(setdiff(result$df$snp, dropsnps)) <= 1)
+        message("dropping ",length(dropsnps),"/",nrow(obj$results)," : ",length(drop1)," + ",length(drop2))
+        if(length(setdiff(obj$results$snp, dropsnps)) <= 1)
             return(NULL)
-        df <- copy(result$df)
-        df3 <- copy(result$df3)
+        df <- copy(obj$results)
+        df3 <- copy(obj$results.H3)
         df[snp %in% drop1, c("lbf1","lbf4"):=list(0,0)]
         df[snp %in% drop2, c("lbf2","lbf4"):=list(0,0)]
+        ## PP | H4
+        my.denom.log.abf <- logsum(df$lbf4)
+        newresult[[ paste0("SNP.PP.H4.row",r) ]] <- exp(df$lbf4 - my.denom.log.abf)
+        newresult[[ paste0("z.df1.row",r) ]] <- ifelse(df$snp %in% drop1, 0, df$z.df1)
+        newresult[[ paste0("z.df2.row",r) ]] <- ifelse(df$snp %in% drop2, 0, df$z.df2)
+
         ## df[snp %in% drop2, c("lbf2","lbf4"):=list(log(1/1000),log(1/1000))]
         ## df3[snp1 %in% drop1, lABF.df1:=0]
         ## df3[snp2 %in% drop2, lABF.df2:=0]
@@ -366,26 +388,16 @@ coloc.process <- function(result,hits1=NULL,hits2=NULL,LD=NULL,r2thr=0.01,p1=1e-
         ## mask <- masks[[ todo[r,1] ]], mask2[[ todo[r,2] ]]),"")
         ## friends <-  apply(abs(LD[masks[,,drop=FALSE]) > sqrt(r2thr),2,any)
         ## dropsnps <- rownames(LD)[which(friends)]
-        ## cbind(.f(result$df[ !(snp %in% dropsnps), ],
-        ##          result$df3[ !(snp1 %in% dropsnps | snp2 %in% dropsnps), ]),
+        ## cbind(.f(obj$results[ !(snp %in% dropsnps), ],
+        ##          obj$results3[ !(snp1 %in% dropsnps | snp2 %in% dropsnps), ]),
         tmp <- cbind(.f(df, df3))
         tmp$hit1=hits1[todo[1,r] ]
         tmp$hit2=hits2[ todo[2,r] ]
-        ## zz <- signif(coloc.res[2 : 6], 2)
-	## 			ii <- zz < 0.01
-	## 			zz[ii] <- "< 0.01"
-	## 			zz[!ii] <- paste0("= ", zz[!ii])		
-	## 			tt <- paste(paste0("H", 0 : 4, " ", zz), collapse = ", ")
-						
-        ## DF <- melt(df, measure.vars = c("trait", "eqtl"), value.name = "pval")
-        ## DF[, dropped := snp %in% dropsnps]
-        ## dummy <- rbind(DF[snp %in% hits1 & variable == 'trait', ], DF[snp %in% hits2 & variable == 'eqtl', ]) ​
-        ## gg <- ggplot(data = DF, aes(x = pos, y = -log10(pval), colour = dropped)) + geom_point(data = DF, aes(alpha = 0.5)) + geom_vline(data = dummy, aes(xintercept = pos, colour = dropped), linetype = "dashed") + facet_wrap(variable ~ ., nrow = 2, scales = "free_y") + scale_colour_manual(values = c("firebrick1", "grey50")) + theme(legend.position="none") + labs(title = tt)
-        ## gg
-        tmp
-    })  %>% rbindlist()
-   ret 
-    ## c(result,ret)
+        ret[[i]] <- copy(tmp)
+    }
+    return(list(summary=rbindlist(ret),
+                results=newresult,
+                priors=c(p1=p1,p2=p2,p12=p12)))
 }
 
 ##' Convert binomial to linear regression
@@ -426,7 +438,7 @@ bin2lin <- function(D) {
 est_all_cond <- function(D,FM) {
     if(D$type=="cc")
         D <- bin2lin(D)
-    x <- as.data.table(D[c("beta","varbeta","snp","MAF")])
+    x <- as.data.table(D[intersect(c("beta","varbeta","snp","MAF","position"),names(D))])
     x[,z:=beta/sqrt(varbeta)]
     if(length(FM)==1)
         return(structure(list(x),names=names(FM)))
@@ -438,12 +450,19 @@ est_all_cond <- function(D,FM) {
     cond <- lapply(seq_along(FM), function(i) {
         tmp <- est_cond(x,D$LD,D$N,YY=YY,
                         sigsnps=names(FM)[-i])
-        tmp$MAF <- x$MAF[ match(tmp$snp, x$snp) ] 
+        ## tmp$MAF <- x$MAF[ match(tmp$snp, x$snp) ]
+        ## if("position" %in% names(D))
+        ##     tmp$position <- x$position[ match(tmp$snp, x$snp) ]
         tmp
     })
     names(cond) <- names(FM)
     cond
 }
+
+## merge_cond <- function(cond,D) {
+##     m <- match(cond$snp,D$snp)
+##     for(nm in intersect(names(D), c("MAF","position")))
+        
 
 ##' New coloc function, builds on coloc.abf() by allowing for multiple
 ##' independent causal variants per trait through conditioning or
@@ -484,9 +503,8 @@ coloc.signals <- function(dataset1, dataset2,
         dataset1$MAF <- MAF
     if(!("MAF" %in% names(dataset2)) & !is.null(MAF))
         dataset2$MAF <- MAF
-    if(!("LD" %in% names(dataset1)) & !is.null(LD)) {
+    if(!("LD" %in% names(dataset1)) & !is.null(LD)) 
         dataset1$LD <- LD
-    }
     if(!("LD" %in% names(dataset2)) & !is.null(LD))
         dataset2$LD <- LD
     if(!("method" %in% names(dataset1)) & !is.null(method))
@@ -522,11 +540,11 @@ coloc.signals <- function(dataset1, dataset2,
     ## conditionals if needed
     if(!is.null(fm1) && dataset1$method=="cond") {
         cond1 <- est_all_cond(dataset1,fm1)
-        X1 <- dataset1[ setdiff(names(dataset1),names(cond1[[1]])) ]
+        X1 <- dataset1[ intersect(names(dataset1), c("N","sdY","type")) ]
     }
     if(!is.null(fm2) && dataset2$method=="cond")  {
         cond2 <- est_all_cond(dataset2,fm2)
-        X2 <- dataset2[ setdiff(names(dataset2),names(cond2[[1]])) ]
+        X2 <- dataset2[ intersect(names(dataset1), c("N","sdY","type")) ]
     }
 
     ## double mask
@@ -541,15 +559,14 @@ coloc.signals <- function(dataset1, dataset2,
         todo <- expand.grid(i=seq_along(cond1),j=seq_along(cond2))
         res <- vector("list",nrow(todo))
         for(k in 1:nrow(todo)) {
-            col <- coloc.detail(c(cond1[[ todo[k,"i"] ]],X1),
-                                c(cond2[[ todo[k,"j"] ]],X2),
+            col <- coloc.detail(dataset1=c(cond1[[ todo[k,"i"] ]],X1),
+                                dataset2=c(cond2[[ todo[k,"j"] ]],X2),
                                 p1=p1,p2=p2,p12=p12)
             res[[k]] <- coloc.process(col,
                                  hits1=names(fm1)[ todo[k,"i"] ],
                                  hits2=names(fm2)[ todo[k,"j"] ],
                                  LD1=dataset1$LD, LD2=dataset2$LD) 
         }
-        res <- rbindlist(res)
     }
 
     ## double -
@@ -561,7 +578,7 @@ coloc.signals <- function(dataset1, dataset2,
                              hits1=names(fm1)[ 1 ],
                              hits2=names(fm2)[1],
                              LD1=dataset1$LD, LD2=dataset2$LD,r2thr=r2thr)
-    }
+  }
         
     ## cond-
     if(dataset1$method=="cond" && dataset2$method=="") {
@@ -575,7 +592,6 @@ coloc.signals <- function(dataset1, dataset2,
                                  hits2=names(fm2)[1],
                                  LD1=dataset1$LD, LD2=dataset2$LD,r2thr=r2thr) #, ...)
         }
-        res <- rbindlist(res)
     }
 
     ## cond-mask
@@ -590,7 +606,6 @@ coloc.signals <- function(dataset1, dataset2,
                                  hits2=names(fm2),
                                  LD1=dataset1$LD, LD2=dataset2$LD,r2thr=r2thr) #, ...)
         }
-        res <- rbindlist(res)
     }
 
     ## -cond
@@ -623,12 +638,38 @@ coloc.signals <- function(dataset1, dataset2,
         res <- rbindlist(res)
     }
 
-    p1 <- data.table(hit1=names(fm1),hit1.margz=c(fm1))
-    res <- merge(res,p1,by="hit1")
-    p2 <- data.table(hit2=names(fm2),hit2.margz=c(fm2))
-    res <- merge(res,p2,by="hit2")
+    if(dataset1$method=="cond" || dataset2$method=="cond") {
+        summary <- rbindlist(lapply(res,"[[","summary"))
+        results <- c(list(res[[1]]$results),
+                     lapply(res[-1], function(x) x$results[,grep("SNP.PP.H4|z.df1|z.df2",names(x$results)),drop=FALSE,with=FALSE]))
+        results <- do.call("cbind",results)
+        ppvars <- grep("SNP.PP.H4",names(results))
+        names(results)[ppvars] <- paste0("SNP.PP.H4.row",seq_along(ppvars))
+        ppvars <- grep("z.df1",names(results))
+        names(results)[ppvars] <- paste0("z.df1.row",seq_along(ppvars))
+        ppvars <- grep("z.df2",names(results))
+        names(results)[ppvars] <- paste0("z.df2.row",seq_along(ppvars))
+    } else {
+        summary <- res[["summary"]]
+        results <- res[["results"]]
+    }
+
+    ## ## if cond-cond, will have lost position
+    ## if(dataset1$method=="cond" && dataset2$method=="cond" &&
+    ##    ("position" %in% names(dataset1))) {
+    ##     m <- match(results$snp,dataset1$position)
+    ##     results <- cbind(snp=results$snp, results[,-1],
+    ##                      position=dataset1$position[m])
+    ## }
+
+    d1 <- data.table(hit1=names(fm1),hit1.margz=c(fm1))
+    res <- merge(summary,d1,by="hit1")
+    d2 <- data.table(hit2=names(fm2),hit2.margz=c(fm2))
+    res <- merge(res,d2,by="hit2")
     
-    res 
+    list(summary=res,
+         results=results,
+         priors=c(p1=p2,p2=p2,p12=p12))
 }